rename neofinrl to finrl meta
