+++ finrl/drl_agents/stablebaselines3/models.py
@@ -5,8 +5,8 @@ import time
 import numpy as np
 import pandas as pd
 from finrl.apps import config
-from finrl.neo_finrl.env_stock_trading.env_stocktrading import StockTradingEnv
-from finrl.neo_finrl.preprocessor.preprocessors import data_split
+from finrl.finrl_meta.env_stock_trading.env_stocktrading import StockTradingEnv
+from finrl.finrl_meta.preprocessor.preprocessors import data_split
 from stable_baselines3 import A2C, DDPG, PPO, SAC, TD3
 from stable_baselines3.common.callbacks import BaseCallback
 from stable_baselines3.common.noise import (
