+++ finrl/finrl_meta/data_processor.py
@@ -1,8 +1,8 @@
 import numpy as np
 import pandas as pd
-from finrl.neo_finrl.data_processors.processor_alpaca import AlpacaProcessor as Alpaca
-from finrl.neo_finrl.data_processors.processor_wrds import WrdsProcessor as Wrds
-from finrl.neo_finrl.data_processors.processor_yahoofinance import (
+from finrl.finrl_meta.data_processors.processor_alpaca import AlpacaProcessor as Alpaca
+from finrl.finrl_meta.data_processors.processor_wrds import WrdsProcessor as Wrds
+from finrl.finrl_metas.data_processors.processor_yahoofinance import (
     YahooFinanceProcessor as YahooFinance,
 )
