+++ finrl/finrl_meta/preprocessor/preprocessors.py
@@ -3,7 +3,7 @@ import datetime
 import numpy as np
 import pandas as pd
 from finrl.apps import config
-from finrl.neo_finrl.preprocessor.yahoodownloader import YahooDownloader
+from finrl.finrl_meta.preprocessor.yahoodownloader import YahooDownloader
 from stockstats import StockDataFrame as Sdf
